Add better struct for the components
