+++ src/stories/containers/RecognizedDelegates/TotalAndKeyStatsComponent/TotalAndkeyStatusComponent.tsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled';
 import React from 'react';
 import KeyStats from '../KeyStats/KeyStats';
-import Range from './Range';
-import TotalDaiExpense from './TotalDaiExpense';
+import Range from '../components/Range';
+import TotalDaiExpense from '../components/TotalDaiExpense';
 
 interface Props {
   totalDAI: string;
