+++ src/stories/containers/RecognizedDelegates/components/DelegateBarPercentTotal.tsx
