+++ src/stories/containers/RecognizedDelegates/components/GenericDelegateCard.tsx
