+++ src/stories/containers/RecognizedDelegates/components/LegendItem.tsx
