+++ src/stories/containers/RecognizedDelegates/components/PercentageTotalCard.tsx
@@ -1,23 +1,23 @@
 import styled from '@emotion/styled';
 import React from 'react';
-import GenericDelegateCard from '../../GenericDelegateCard';
+import KeyStatsCard from './GenericDelegateCard';
 import { DescriptionDelegates } from './TotalRecognizedDelegatesCard';
 
 interface Props {
-  annual: string;
+  percent: string;
 }
 
-const MedianAnnualDai: React.FC<Props> = ({ annual }) => (
+const PercentageTotalCard: React.FC<Props> = ({ percent }) => (
   <ExtendedGenericDelegate>
-    <Annual>{`${annual} dai`}</Annual>
-    <DescriptionDelegates>{'Median Annual Compensation / Delegate'}</DescriptionDelegates>
+    <Annual>{percent}</Annual>
+    <DescriptionDelegatesExtended>{'Percentage of Total DAO Expense Nov 2021 - Jun 2023'}</DescriptionDelegatesExtended>
   </ExtendedGenericDelegate>
 );
 
-export default MedianAnnualDai;
+export default PercentageTotalCard;
 
-const ExtendedGenericDelegate = styled(GenericDelegateCard)({
-  padding: '8px 26px',
+const ExtendedGenericDelegate = styled(KeyStatsCard)({
+  padding: '8px 18.5px',
   display: 'flex',
   flexDirection: 'column',
   alignItems: 'center',
@@ -35,4 +35,8 @@ export const Annual = styled.div({
   color: '#243465',
   marginBottom: 4,
   textTransform: 'uppercase',
+});
+
+const DescriptionDelegatesExtended = styled(DescriptionDelegates)({
+  width: 110,
 });
