+++ src/stories/containers/RecognizedDelegates/components/ShadowDelegatesCard.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import React from 'react';
-import GenericDelegateCard from '../../GenericDelegateCard';
+import GenericDelegateCard from './GenericDelegateCard';
 import { DescriptionDelegates, TotalDelegates } from './TotalRecognizedDelegatesCard';
 
 interface Props {
